update reports
Implemented report creation to occur only during the first run, with subsequent updates handled in real-time (Added functionality for canceling purchases and buying multi-entry tickets, with plans to incorporate complaint handling and user purchases after merging the code). Updated the month selection in the combo box to dynamically adjust based on the selected year, preventing the choice of future months that are not yet available.